casi terminado el editor

+++ editor/editor.h
@@ -31,8 +31,8 @@ class Editor{
     SocketPrefijo *receptor;
     Ventana *ventana;
     std::string nombre_nivel;
-    size_t columnas;
-    size_t filas;
+    unsigned int columnas;
+    unsigned int filas;
     SDL libreria;
 };
